Fixed menu animation

+++ src/menu.hpp
@@ -63,6 +63,6 @@ private:
     sf::SoundBuffer buffer_; // Buffer for the button sample
     sf::Text sIP_; // Holds the IP address to be joined
     sf::Clock clock_; // Timer for the menu animation
-    int x_; // Variable for menu animation position change
+    float x_; // Variable for menu animation position change
     int menu_status; // Indicates which menu scenario is to be loaded
 };
